move seriExp path
