+++ dubbo-demo/dubbo-demo-spring-boot/dubbo-demo-spring-boot-consumer/src/main/java/org/apache/dubbo/springboot/demo/consumer/ConsumerApplication.java
@@ -20,6 +20,7 @@ package org.apache.dubbo.springboot.demo.consumer;
 import org.apache.dubbo.config.annotation.DubboReference;
 import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
 import org.apache.dubbo.springboot.demo.DemoService;
+import org.apache.dubbo.springboot.demo.Person;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -42,6 +43,6 @@ public class ConsumerApplication {
     }
 
     public String doSayHello(String name) {
-        return demoService.sayHello(name);
+        return demoService.sayHello(new Person(10, name));
     }
 }
