+++ dubbo-demo/dubbo-demo-spring-boot/dubbo-demo-spring-boot-interface/src/main/java/org/apache/dubbo/springboot/demo/DemoService.java
@@ -16,14 +16,12 @@
  */
 package org.apache.dubbo.springboot.demo;
 
-import java.util.concurrent.CompletableFuture;
-
 public interface DemoService {
 
-    String sayHello(String name);
+    String sayHello(Person person);
 
-    default CompletableFuture<String> sayHelloAsync(String name) {
-        return CompletableFuture.completedFuture(sayHello(name));
-    }
+//    default CompletableFuture<String> sayHelloAsync(String name) {
+//        return CompletableFuture.completedFuture(sayHello(name));
+//    }
 
 }
